undo redo refactored
